Improve stats consistency
There is some fluctiation between the different stats datasets
due to the nature of the live voting. This causes some weirdness
in the reported statistics which can be misleading for the user.

In order to improve the consistency, we consider the start time of
the data aggregation which is used for all datasets. Therefore,
all datasets will process exactly the same votes causing
consistent results.